requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
Django==4.1
django-ckeditor==6.5.0
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-decouple==2.1
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
jdatetime==4.1.0
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.2
